server/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
Pillow==10.3.0
stripe==9.8.0
django-environ==0.11.2
setuptools==70.0.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7